+++ config/eslint-comments.js
@@ -1,0 +1,10 @@
+'use strict';
+
+module.exports = {
+	plugins: [
+		'eslint-comments',
+	],
+	extends: [
+		'../rules/plugin-eslint-comments',
+	].map(require.resolve),
+};