+++ rules/plugin-react-hooks.js
@@ -1,0 +1,10 @@
+'use strict';
+
+// https://mysticatea.github.io/eslint-plugin-eslint-comments/
+
+module.exports = {
+	rules: {
+		'react-hooks/rules-of-hooks': 'error',
+		'react-hooks/exhaustive-deps': 'error',
+	},
+};